+++ packages/camera/camera_android_camerax/test/android_camera_camerax_test.mocks.dart
@@ -458,6 +458,15 @@ class MockCameraInfo extends _i1.Mock implements _i3.CameraInfo {
             returnValueForMissingStub: _FakeCameraInfo_8(this, Invocation.method(#pigeon_copy, [])),
           )
           as _i3.CameraInfo);
+
+  @override
+  _i5.Future<bool> isZslSupported() =>
+      (super.noSuchMethod(
+            Invocation.method(#isZslSupported, []),
+            returnValue: _i5.Future<bool>.value(false),
+            returnValueForMissingStub: _i5.Future<bool>.value(false),
+          )
+          as _i5.Future<bool>);
 }
 
 /// A class which mocks [CameraCharacteristicsKey].