+++ packages/camera/camera_avfoundation/darwin/Tests/Mocks/MockCapturePhotoOutput.swift
@@ -13,6 +13,8 @@ final class MockCapturePhotoOutput: NSObject, CapturePhotoOutput {
   var availablePhotoCodecTypes: [AVVideoCodecType] = []
   var isHighResolutionCaptureEnabled = false
   var supportedFlashModes: [AVCaptureDevice.FlashMode] = []
+  var flutterZeroShutterLagSupported = false
+  var flutterZeroShutterLagEnabled = false
 
   // Stub that is called when the corresponding public method is called.
   var capturePhotoWithSettingsStub: